+++ modules/config/rices.nix
@@ -1,0 +1,17 @@
+{delib, ...}:
+delib.module {
+  name = "rices";
+
+  options = with delib; let
+    rice = {
+      options = riceSubmoduleOptions;
+    };
+  in {
+    rice = riceOption rice;
+    rices = ricesOption rice;
+  };
+
+  home.always = {myconfig, ...}: {
+    assertions = delib.riceNamesAssertions myconfig.rices;
+  };
+}
