+++ modules/features/audio.nix
@@ -1,0 +1,33 @@
+{
+  delib,
+  pkgs,
+  ...
+}: let
+  inherit (delib) singleEnableOption;
+in
+  delib.module {
+    name = "features.audio";
+
+    options = singleEnableOption false;
+
+    nixos.ifEnabled = {myconfig, ...}: {
+      environment.systemPackages = with pkgs; [
+        ## Audio
+        alsa-oss
+        alsa-utils
+        alsa-lib
+        sox
+      ];
+
+      services = {
+        ## Audio
+        pipewire = {
+          enable = true;
+          alsa.enable = true;
+          alsa.support32Bit = true;
+          pulse.enable = true;
+          jack.enable = true;
+        };
+      };
+    };
+  }
