+++ modules/features/bluetooth.nix
@@ -1,0 +1,36 @@
+{
+  delib,
+  pkgs,
+  ...
+}: let
+  inherit (delib) singleEnableOption;
+in
+  delib.module {
+    name = "features.bluetooth";
+
+    options = singleEnableOption false;
+
+    nixos.ifEnabled = {myconfig, ...}: {
+      environment.systemPackages = with pkgs; [
+        blueman
+      ];
+      hardware = {
+        bluetooth = {
+          enable = true;
+          powerOnBoot = true;
+          settings = {
+            General = {
+              Name = "Hello";
+              ControllerMode = "dual";
+              FastConnectable = "true";
+              Experimental = "true";
+              Enable = "Source,Sink,Media,Socket";
+            };
+            Policy = {
+              AutoEnable = "true";
+            };
+          };
+        };
+      };
+    };
+  }
