+++ modules/features/darknet.nix
@@ -1,0 +1,19 @@
+{
+  delib,
+  pkgs,
+  ...
+}: let
+  inherit (delib) singleEnableOption;
+in
+  delib.module {
+    name = "features.darknet";
+
+    options = singleEnableOption false;
+
+    nixos.ifEnabled = {myconfig, ...}: {
+      services = {
+        tailscale.enable = true;
+        fail2ban.enable = true;
+      };
+    };
+  }
