+++ modules/features/secrets.nix
@@ -1,0 +1,19 @@
+{
+  delib,
+  pkgs,
+  ...
+}: let
+  inherit (delib) singleEnableOption;
+in
+  delib.module {
+    name = "features.secrets";
+
+    options = singleEnableOption false;
+
+    nixos.ifEnabled = {myconfig, ...}: {
+      environment.systemPackages = with pkgs; [
+        # sops
+        proton-pass
+      ];
+    };
+  }
