--- nix/home-manager/default.nix
@@ -1,112 +1,0 @@
-{
-  lib,
-  pkgs,
-  stateVersion,
-  username,
-  ...
-}: let
-  inherit (pkgs.stdenv) isDarwin isLinux;
-in {
-  home = {
-    inherit stateVersion;
-    inherit username;
-    homeDirectory =
-      if isDarwin
-      then "/Users/${username}"
-      else "/home/${username}";
-
-    packages = with pkgs;
-      [
-        cpufetch
-        fastfetch
-        ipfetch
-        onefetch
-      ]
-      ++ lib.optionals isLinux [
-        ramfetch
-      ]
-      ++ lib.optionals isDarwin [
-        m-cli
-      ];
-    sessionVariables = {
-      EDITOR = "nvim";
-      SYSTEMD_EDITOR = "nvim";
-      VISUAL = "nvim";
-    };
-  };
-
-  gtk =
-    if isLinux
-    then {
-      enable = true;
-      theme = {
-        name = lib.mkForce "adw-gtk3-dark";
-        package = lib.mkForce pkgs.adw-gtk3;
-      };
-      iconTheme.package = pkgs.papirus-icon-theme;
-      iconTheme.name = "Papirus";
-
-      gtk3.extraConfig = {
-        Settings = ''
-          gtk-application-prefer-dark-theme=1
-        '';
-      };
-
-      gtk4.extraConfig = {
-        Settings = ''
-          gtk-application-prefer-dark-theme=1
-        '';
-      };
-    }
-    else {};
-
-  services.mpris-proxy.enable =
-    if isLinux
-    then true
-    else false;
-
-  # Workaround home-manager bug
-  # - https://github.com/nix-community/home-manager/issues/2033
-  news = {
-    display = "silent";
-    entries = lib.mkForce [];
-  };
-
-  nix = {
-    package = pkgs.nixVersions.latest;
-  };
-
-  programs = {
-    zsh.enable = true;
-    zed-editor = {
-      enable = true;
-      extensions = [
-        "nix"
-        "dockerfile"
-        "toml"
-        "html"
-        "templ"
-        "latex"
-        "svelte"
-        "golangci-lint"
-        "astro"
-        "python-lsp"
-        "ocaml"
-        "vhdl"
-        "verilog"
-      ];
-      userSettings = {
-        ui_font_size = 16;
-        buffer_font_size = 16;
-        telemetry.enable = false;
-        vim_mode = true;
-        theme = {
-          mode = "dark";
-          dark = "One Dark";
-          light = "One Light";
-        };
-      };
-    };
-    home-manager.enable = true;
-  };
-}
