--- nix/home-manager/utils/default.nix
@@ -1,13 +1,0 @@
-{
-  inputs,
-  stateVersion,
-  ...
-}: let
-  helpers = import ./helpers.nix {inherit inputs stateVersion;};
-in {
-  inherit
-    (helpers)
-    mkHome
-    forAllSystems
-    ;
-}
