--- nix/lib/module/default.nix
@@ -1,26 +1,0 @@
-{
-  lib,
-  inputs,
-  namespace,
-  snowfall-inputs,
-}:
-with lib; rec {
-  mkOpt = type: default: description:
-    mkOption {inherit type default description;};
-
-  mkOpt' = type: default: mkOpt type default null;
-
-  mkBoolOpt = mkOpt types.bool;
-
-  mkBoolOpt' = mkOpt' types.bool;
-
-  enabled = {
-    enable = true;
-  };
-
-  disabled = {
-    enable = false;
-  };
-
-  nullOrEnums = enums: with types; nullOr (enum enums);
-}
