--- nix/modules/nixos/power-efficient/default.nix
@@ -1,22 +1,0 @@
-{
-  pkgs,
-  config,
-  lib,
-  namespace,
-  ...
-}:
-with lib; let
-  cfg = config.${namespace}.hardware.power-efficient;
-in {
-  options.${namespace}.hardware.power-efficient = with types; {
-    enable = mkEnableOption "Audio support";
-  };
-
-  config = mkIf cfg.enable {
-    hardware = {};
-    services = {
-      tlp.enable = true;
-      power-profiles-daemon.enable = false;
-    };
-  };
-}
