--- nix/packages/catls/default.nix
@@ -1,7 +1,0 @@
-{pkgs, ...}:
-pkgs.writers.writeRubyBin "catls" {
-  libraries = [
-  ];
-} ''
-  ${builtins.readFile ./catls.rb}
-''
