--- nix/packages/convert_img/default.nix
@@ -1,8 +1,0 @@
-{
-  pkgs,
-  namespace,
-  ...
-}:
-pkgs.writeShellScriptBin "convert_img" ''
-  ${pkgs."${namespace}".python-venv}/bin/python ${./convert_img.py} $@
-''
