--- nix/packages/python-venv/default.nix
@@ -1,21 +1,0 @@
-{pkgs, ...}:
-pkgs.python312.withPackages (
-  ps:
-    with ps; [
-      numpy
-      requests
-      pandas
-      scipy
-      matplotlib
-      scikitlearn
-      pyarrow
-      black
-      requests
-      mypy
-      beautifulsoup4
-      pillow
-      pypdf
-      pip
-      sympy
-    ]
-)
