--- nix/shells/default/scripts.nix
@@ -1,16 +1,0 @@
-# Scripts for development shell
-{pkgs, ...}: {
-  dx = {
-    exec = ''$EDITOR $REPO_ROOT/flake.nix'';
-    description = "Edit the flake.nix";
-  };
-  lint = {
-    exec = ''
-      export REPO_ROOT=$(git rev-parse --show-toplevel)
-      ${pkgs.statix}/bin/statix check $REPO_ROOT/flake.nix
-      ${pkgs.deadnix}/bin/deadnix $REPO_ROOT/flake.nix
-      nix flake check --all-systems
-    '';
-    description = "Run linters";
-  };
-}
