+++ templates/go-shell/.envrc
