+++ templates/go-shell/.gitignore
