+++ templates/rust-shell/.envrc
