+++ templates/rust-shell/.gitignore
