+++ templates/rust-shell/flake.nix
